# server/openapi/v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Streamarr Core Server API",
    "description": "Interface-agnostic API for Usenet search, resolve, and byte-range streaming (BRIEF §6.2).",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "episode",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "imdbId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tmdbId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "profileId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/search": {
      "post": {
        "tags": [
          "Search"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebugSearchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DebugSearchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DebugSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DebugSearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugSearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tv/search": {
      "get": {
        "tags": [
          "TvCatalog"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TvSeriesSearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvSeriesSearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvSeriesSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DebugSearchRequest": {
        "required": [
          "q"
        ],
        "type": "object",
        "properties": {
          "q": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "season": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "episode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "imdbId": {
            "type": "string",
            "nullable": true
          },
          "tmdbId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "profileId": {
            "type": "string",
            "nullable": true
          },
          "profile": {
            "$ref": "#/components/schemas/QualityProfile"
          }
        },
        "additionalProperties": false,
        "description": "Request body of POST /api/v1/debug/search (BRIEF §6.2). Mirrors the /search params."
      },
      "DebugSearchResponse": {
        "required": [
          "indexers",
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebugWorkDto"
            },
            "nullable": true
          },
          "indexers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexerDiagnosticDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Response of POST /api/v1/debug/search: every release including rejected ones, with\r\nparsed fields, per-rule score breakdown and rejection reasons (BRIEF §6.2). Powers\r\nthe ranker-tuning view; still never exposes NZB URLs."
      },
...
      "ErrorDetail": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "host": {
            "type": "string",
            "description": "Populated only for the `nzb_host_not_allowed` error: the rejected download host.",
            "nullable": true
          },
          "indexerId": {
            "type": "string",
            "description": "Populated only for the `nzb_host_not_allowed` error: the owning indexer's id.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "description": "Typed error envelope rendered consistently by every endpoint."
      },
...
      "IndexerDiagnosticDto": {
        "required": [
          "indexerId",
          "indexerName",
          "status"
        ],
        "type": "object",
        "properties": {
          "indexerId": {
            "type": "string",
            "nullable": true
          },
          "indexerName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "itemCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "elapsedMs": {
            "type": "number",
            "format": "double"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Per-indexer fan-out diagnostics for the debug view."
      },
...
      "QualityProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "appliesTo": {
            "type": "string",
            "nullable": true
          },
          "importedFrom": {
            "type": "string",
            "nullable": true
          },
          "importedProfileId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "importedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "preferredResolutions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "preferredSources": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "preferredCodecs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "preferredLanguages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "groupAllowList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "groupDenyList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "resolutionWeight": {
            "type": "integer",
            "format": "int32"
          },
          "sourceWeight": {
            "type": "integer",
            "format": "int32"
          },
          "codecWeight": {
            "type": "integer",
            "format": "int32"
          },
          "languageWeight": {
            "type": "integer",
            "format": "int32"
          },
          "audioWeight": {
            "type": "integer",
            "format": "int32"
          },
          "sizeWeight": {
            "type": "integer",
            "format": "int32"
          },
          "properRepackBonus": {
            "type": "integer",
            "format": "int32"
          },
          "recencyBonus": {
            "type": "integer",
            "format": "int32"
          },
          "grabsBonus": {
            "type": "integer",
            "format": "int32"
          },
          "groupAllowBonus": {
            "type": "integer",
            "format": "int32"
          },
          "groupDenyPenalty": {
            "type": "integer",
            "format": "int32"
          },
          "customFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatScore"
            },
            "nullable": true
          },
          "minimumCustomFormatScore": {
            "type": "integer",
            "format": "int32"
          },
          "minBytesPerMinute": {
            "type": "integer",
            "format": "int64"
          },
          "maxBytesPerMinute": {
            "type": "integer",
            "format": "int64"
          },
          "sizeBands": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SizeBand"
            },
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SearchResponse": {
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Response of GET /api/v1/search — only works with at least one accepted release;\r\nrejected releases remain available exclusively through the debug contract."
      },
...
      "TvSeriesSearchResponse": {
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvSeriesDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "TMDB-ranked TV series candidates. The endpoint is intentionally capped at three."
      },
      "WorkDto": {
        "required": [
          "mediaType",
          "releases",
          "title",
          "workId"
        ],
        "type": "object",
        "properties": {
          "workId": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "description": "\"movie\" or \"tv\".",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tmdbId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "imdbId": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "posterUrl": {
            "type": "string",
            "nullable": true
          },
          "backdropUrl": {
            "type": "string",
            "nullable": true
          },
          "runtimeMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "originalTitle": {
            "type": "string",
            "nullable": true
          },
          "tagline": {
            "type": "string",
            "nullable": true
          },
          "officialRating": {
            "type": "string",
            "nullable": true
          },
          "communityRating": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "studios": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "productionLocations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbPerson"
            },
            "nullable": true
          },
          "trailerUrl": {
            "type": "string",
            "nullable": true
          },
          "addStreamarrBadge": {
            "type": "boolean"
          },
          "season": {
            "type": "integer",
            "description": "Set for TV works.",
            "format": "int32",
            "nullable": true
          },
          "episode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "releases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "One aggregated work with its ranked releases (BRIEF §6.2 / §7.4)."
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Machine API key or admin session JWT.",
        "scheme": "bearer",
        "bearerFormat": "API key or JWT"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}